ISSUE #1082 - Move users list to new component
