+++ frontend/components/account/pug/account-user-management.pug
@@ -42,74 +42,11 @@
 			md-selected="vm.selectedTab"
 			md-on-select="vm.onTabChange()") 
 			md-tab(label="Users")
-				.usersList
-					.usersListHeader(layout="row" layout-align="start center")
-						.usersListCell(layout="row" layout-align="start center" flex="30")
-							list-sort-button(
-								ng-class="{'active': vm.currentSort.type === vm.SORT_TYPES.USERS}"
-								on-change="vm.setSortType(vm.SORT_TYPES.USERS, order)")
-							.usersListCellLabel User
-
-						.usersListCell(layout="row" layout-align="start center" flex="25")
-							list-sort-button(
-								ng-class="{'active': vm.currentSort.type === vm.SORT_TYPES.JOBS}"
-								on-change="vm.setSortType(vm.SORT_TYPES.JOBS, order)")
-							.usersListCellLabel Job
-
-						.usersListCell(layout="row" layout-align="start center" flex="30")
-							list-sort-button(
-								ng-class="{'active': vm.currentSort.type === vm.SORT_TYPES.PERMISSIONS}"
-								on-change="vm.setSortType(vm.SORT_TYPES.PERMISSIONS, order)")
-							.usersListCellLabel Permissions
-
-					.usersListBody
-						div.usersListRow(
-							ng-repeat="member in vm.processedMembers track by member.user"
-							layout="row"
-							layout-align="start center")
-
-							//User details
-							member-card.usersListCell(data="member" flex="30")
-
-							//Job
-							.usersListCell(layout="row" layout-align="start center" flex="25")
-								md-input-container(md-no-float)
-									md-select(
-										ng-model="member.job"
-										placeholder="Select a job"
-										ng-change="vm.onJobChange(member)")
-										md-option(ng-value="" ng-if="vm.jobs.length && member.job")
-											div(layout="row" layout-align="start center")
-												span.usersListJobColor(ng-style="{'background-color': job.color}")
-												.usersListJobName Unassign
-										md-option(
-											ng-value="job._id"
-											ng-repeat="job in vm.jobs track by job._id")
-											div(layout="row" layout-align="start center")
-												span.usersListJobColor(ng-style="{'background-color': job.color}")
-												.usersListJobName {{job._id}}
-												
-
-							//Permissions
-							.usersListCell(layout="row" layout-align="start center" flex="30")
-								md-input-container(md-no-float)
-									md-select(
-										ng-model="member.isAdmin"
-										ng-disabled="member.isCurrentUser"
-										placeholder="Select a permission"
-										ng-change="vm.onPermissionsChange(member)")
-										md-option(
-											ng-value="value.isAdmin"
-											ng-repeat="value in vm.TEAMSPACE_PERMISSIONS")
-											div(layout="row" layout-align="start center") {{value.label}}
-
-							.flex
-
-							//Remove action
-							md-button.md-icon-button.usersListRemoveButton(
-								ng-if="!member.isCurrentUser"
-								ng-click="vm.removeMember(member)")
-								md-icon.angular-material-icons remove_circle
+				users-list(
+					members="vm.members"
+					jobs="vm.jobs"
+					current-teamspace="vm.currentTeamspace"
+					on-change="vm.onMembersChange(updatedMembers)")
 				.accountUserManagementFooter(
 					layout="row"
 					layout-align="end center") You have {{}} out of {{}} licences left
