+++ frontend/constants/sorting.ts
@@ -1,0 +1,10 @@
+export const SORT_TYPES = {
+	USERS: 1,
+	JOBS: 2,
+	PERMISSIONS: 3
+};
+
+export const SORT_ORDER_TYPES = {
+	ASCENDING: "asc",
+	DESCENDING: "desc"
+};
