+++ frontend/constants/teamspace-permissions.ts
@@ -1,0 +1,11 @@
+export const TEAMSPACE_PERMISSIONS = {
+	admin: {
+		isAdmin: true,
+		key: "teamspace_admin",
+		label: "Teamspace admin"
+	},
+	user: {
+		isAdmin: false,
+		label: "User"
+	}
+};
